adjustments to demo data from changed requirements
